+++ inst/include/rmolib/random/multivariate/mdcm_exmo_distribution.hpp
@@ -20,13 +20,13 @@ namespace random {
 template <typename _RealType, typename _ExponentialDistribution,
           typename _UniformIntDistribution, typename _DiscreteDistribution,
           typename _ShuffleOperation>
-class markovian_exmo_distribution {
+class mdcm_exmo_distribution {
  public:
   using result_type = std::vector<_RealType>;
 
   class param_type {
    public:
-    using distribution_type = markovian_exmo_distribution;
+    using distribution_type = mdcm_exmo_distribution;
 
     param_type() { __init_empty(); };
 
@@ -63,7 +63,7 @@ class markovian_exmo_distribution {
       return out;
     }
 
-    friend class markovian_exmo_distribution;
+    friend class mdcm_exmo_distribution;
 
     friend bool operator==(const param_type& lhs, const param_type& rhs) {
       return lhs.dim_ == rhs.dim_ && lhs.markov_parm_ == rhs.markov_parm_;
@@ -144,7 +144,7 @@ class markovian_exmo_distribution {
     void __init(_InputIterator first, _InputIterator last) {
       static_assert(
           type_traits::is_input_iterator_v<_InputIterator>,
-          "Class template rmolib::random::markovian_exmo_distribution<>: "
+          "Class template rmolib::random::mdcm_exmo_distribution<>: "
           "_InputIterator only be initialized with input iterator");
       if (first == last) {
         __init_empty(first, last);
@@ -158,34 +158,34 @@ class markovian_exmo_distribution {
 
     static_assert(
         std::is_floating_point_v<_RealType>,
-        "Class template rmolib::random::markovian_exmo_distribution<> must be "
+        "Class template rmolib::random::mdcm_exmo_distribution<> must be "
         "parametrized with floating point type");
   };
 
-  markovian_exmo_distribution() = default;
+  mdcm_exmo_distribution() = default;
 
   template <typename _ForwardIterator>
-  explicit markovian_exmo_distribution(const std::size_t dim,
+  explicit mdcm_exmo_distribution(const std::size_t dim,
                                        _ForwardIterator first,
                                        _ForwardIterator last)
       : parm_{dim, first, last} {}
 
   template <typename _Container>
-  explicit markovian_exmo_distribution(const std::size_t dim,
+  explicit mdcm_exmo_distribution(const std::size_t dim,
                                        const _Container& ex_intensities)
       : parm_{dim, ex_intensities} {}
 
-  explicit markovian_exmo_distribution(const param_type& parm) : parm_{parm} {}
+  explicit mdcm_exmo_distribution(const param_type& parm) : parm_{parm} {}
 
   // Used for construction from a different specialization
   template <
       typename _ExMOParamType,
       std::enable_if_t<
-          !std::is_convertible_v<_ExMOParamType, markovian_exmo_distribution> &&
+          !std::is_convertible_v<_ExMOParamType, mdcm_exmo_distribution> &&
               !std::is_convertible_v<_ExMOParamType, param_type> &&
               is_exmo_param_type_v<_ExMOParamType>,
           int> = 0>
-  explicit markovian_exmo_distribution(_ExMOParamType&& parm)
+  explicit mdcm_exmo_distribution(_ExMOParamType&& parm)
       : parm_{std::forward<_ExMOParamType>(parm)} {}
 
   // compiler generated ctor and assignment op is sufficient
@@ -229,13 +229,13 @@ class markovian_exmo_distribution {
     shuffle(out.begin(), out.end(), engine, uniform_int_dist_);
   }
 
-  friend bool operator==(const markovian_exmo_distribution& lhs,
-                         const markovian_exmo_distribution& rhs) {
+  friend bool operator==(const mdcm_exmo_distribution& lhs,
+                         const mdcm_exmo_distribution& rhs) {
     return lhs.parm_ == rhs.parm_;
   }
 
-  friend bool operator!=(const markovian_exmo_distribution& lhs,
-                         const markovian_exmo_distribution& rhs) {
+  friend bool operator!=(const mdcm_exmo_distribution& lhs,
+                         const mdcm_exmo_distribution& rhs) {
     return !(lhs == rhs);
   }
 
@@ -262,14 +262,14 @@ class markovian_exmo_distribution {
   static_assert(
       type_traits::is_safe_numeric_cast_v<
           _RealType, typename _ExponentialDistribution::result_type>,
-      "Class template rmolib::random::markovian_exmo_distribution<> must be "
+      "Class template rmolib::random::mdcm_exmo_distribution<> must be "
       "parametrized with exponential_distribution-type with matching "
       "result_type");
 
   static_assert(
       type_traits::is_safe_numeric_cast_v<
           std::size_t, typename _DiscreteDistribution::result_type>,
-      "Class template rmolib::random::markovian_exmo_distribution<> must be "
+      "Class template rmolib::random::mdcm_exmo_distribution<> must be "
       "parametrized with discrete_distribution-type with matching "
       "size_type");
 };
@@ -280,13 +280,13 @@ class markovian_exmo_distribution {
   template <class _CharType, class _Traits, typename _RealType, typename
   _ExponentialDistribution> std::basic_ostream<_CharType, _Traits>&
   operator<<(std::basic_ostream<_CharType, _Traits>& os,
-            markovian_exmo_distribution<_RealType,
+            mdcm_exmo_distribution<_RealType,
   _ExponentialDistribution>& dist);
 
   template <class _CharType, class _Traits, typename _RealType, typename
   _ExponentialDistribution> std::basic_istream<_CharType, _Traits>&
   operator>>(std::basic_istream<_CharType, _Traits>& is,
-             markovian_exmo_distribution<_RealType,
+             mdcm_exmo_distribution<_RealType,
   _ExponentialDistribution>& dist);
 */