+++ tests/testthat/test-rmo_am.R
@@ -4,7 +4,7 @@ n <- 1e2
 ## #### Test implementation for the bivariate case ####
 
 test_that("Arnold model implementation for d = 2", {
-  mockery::stub(rmo_arnold, "Rcpp__rmo_arnold", rtest__rmo_arnold)
+  mockery::stub(rmo_am, "Rcpp__rmo_am", rtest__rmo_am)
 
   ## all equal
   args <- list(
@@ -12,7 +12,7 @@ test_that("Arnold model implementation for d = 2", {
     "intensities" = c(1, 1, 1)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## independence
@@ -21,7 +21,7 @@ test_that("Arnold model implementation for d = 2", {
     "intensities" = intensities_linear(2, scale = 0.7)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## comonotone
@@ -30,7 +30,7 @@ test_that("Arnold model implementation for d = 2", {
     "intensities" = intensities_constant(2, constant = 0.7)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Poisson
@@ -41,7 +41,7 @@ test_that("Arnold model implementation for d = 2", {
     )
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Alpha-Stable
@@ -50,7 +50,7 @@ test_that("Arnold model implementation for d = 2", {
     "intensities" = intensities_alpha_stable(2, alpha = 0.25)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Gamma
@@ -59,7 +59,7 @@ test_that("Arnold model implementation for d = 2", {
     "intensities" = intensities_gamma(2, a = 0.4)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Pareto
@@ -68,7 +68,7 @@ test_that("Arnold model implementation for d = 2", {
     "intensities" = intensities_pareto(2, alpha = 0.4, x0 = 1e-4)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Inverse-Gaussian
@@ -77,7 +77,7 @@ test_that("Arnold model implementation for d = 2", {
     "intensities" = intensities_inverse_gaussian(2, eta = 0.5)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Mixed
@@ -86,7 +86,7 @@ test_that("Arnold model implementation for d = 2", {
     "intensities" = c(0.1, 0.005, 2)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 })
 
@@ -94,7 +94,7 @@ test_that("Arnold model implementation for d = 2", {
 ## ## test implementation against original `R` version ####
 
 test_that("Arnold model implementation for d>2", {
-  mockery::stub(rmo_arnold, "Rcpp__rmo_arnold", rtest__rmo_arnold)
+  mockery::stub(rmo_am, "Rcpp__rmo_am", rtest__rmo_am)
 
   ## dimension parameters
   d1 <- 3
@@ -107,7 +107,7 @@ test_that("Arnold model implementation for d>2", {
     "intensities" = rep(0.5, times = 2^d - 1)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## independence
@@ -116,7 +116,7 @@ test_that("Arnold model implementation for d>2", {
     "intensities" = intensities_linear(d, scale = 0.7)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## comonotone
@@ -125,7 +125,7 @@ test_that("Arnold model implementation for d>2", {
     "intensities" = intensities_constant(d, constant = 0.7)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Poisson
@@ -136,7 +136,7 @@ test_that("Arnold model implementation for d>2", {
     )
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Alpha-stable
@@ -145,7 +145,7 @@ test_that("Arnold model implementation for d>2", {
     "intensities" = intensities_alpha_stable(d, alpha = 0.25)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Gamma
@@ -154,7 +154,7 @@ test_that("Arnold model implementation for d>2", {
     "intensities" = intensities_gamma(d, a = 0.4)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Pareto
@@ -163,7 +163,7 @@ test_that("Arnold model implementation for d>2", {
     "intensities" = intensities_pareto(d, alpha = 0.4, x0 = 1e-4)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Inverse-Gaussian
@@ -172,7 +172,7 @@ test_that("Arnold model implementation for d>2", {
     "intensities" = intensities_inverse_gaussian(d, eta = 0.5)
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 
   ## Hierarchical
@@ -184,6 +184,6 @@ test_that("Arnold model implementation for d>2", {
     )
   )
   expect_equal_rn_generation(
-    rmo_arnold, testutils.rmo::rmo_arnold_naive,
+    rmo_am, testutils.rmo::rmo_am_naive,
     args, n, use_seed)
 })